+++ common/src/main/java/net/mehvahdjukaar/amendments/common/LecternEditMenu.java
@@ -1,6 +1,7 @@
 package net.mehvahdjukaar.amendments.common;
 
 import net.mehvahdjukaar.amendments.reg.ModRegistry;
+import net.mehvahdjukaar.moonlight.api.misc.TileOrEntityTarget;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.Container;
@@ -10,6 +11,9 @@ import net.minecraft.world.inventory.LecternMenu;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.inventory.SimpleContainerData;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.LecternBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.entity.LecternBlockEntity;
 
 import java.util.List;
@@ -19,9 +23,9 @@ public class LecternEditMenu extends LecternMenu {
     private final BlockPos pos;
 
     public static LecternEditMenu of(int id, Inventory playerInventory, FriendlyByteBuf packetBuffer) {
-        BlockPos pos = packetBuffer.readBlockPos();
-        return new LecternEditMenu(id,(LecternBlockEntity) playerInventory.player.level()
-                .getBlockEntity(pos), new SimpleContainerData(1));
+        var tile = TileOrEntityTarget.read(packetBuffer);
+        return new LecternEditMenu(id, tile.getBlockEntityOrThrow(playerInventory.player.level(),
+                BlockEntityType.LECTERN), new SimpleContainerData(1));
     }
 
     public LecternEditMenu(int i, LecternBlockEntity container, ContainerData containerData) {
@@ -34,7 +38,7 @@ public class LecternEditMenu extends LecternMenu {
         return ModRegistry.LECTERN_EDIT_MENU.get();
     }
 
-    public BlockPos getPos(){
+    public BlockPos getPos() {
         return this.pos;
     }